refactor: :truck: rename file from rule to custom_check; fix imports
